requirements.txt
alembic==0.9.6
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
decorator==4.0.11
dill==0.2.6
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
future==0.16.0
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.25.0
google-cloud-vision==0.25.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
GroupyAPI==0.10.0
grpcio==1.4.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.0.0
ply==3.8
praw==3.6.2
prawcore==0.12.0
profanityfilter==2.0.3
protobuf==3.3.0
psycopg2-binary==2.7.4
pyasn1==0.2.3
pyasn1-modules==0.0.9
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
redis==2.10.5
requests==2.9.1
rq==0.8.0
rsa==3.4.2
ruamel.yaml==0.15.37
six==1.10.0
SQLAlchemy==1.1.11
tzlocal==1.5.1
update-checker==0.16
urllib3==1.21.1
visitor==0.1.3
Werkzeug==0.12.2
